renderer_vulkan: Rewrite descriptor set management and various small fixes (#77)
* renderer_vulkan: Remove vulkan prefix in SetObjectName

* renderer_vulkan: Rename renderpass cache to render manager

* It is no longer just a cache

* renderer_vulkan: Rewrite descriptor management

* Switch to batched vkUpdateDescriptorSets from cached descriptor sets with templates

* vk_master_semaphore: Remove waitable atomic

* These are buggy on some platforms and regular condition_variables are faster most of the time

* vk_texture_runtime.cpp: remove outdated references

* vk_render_manager: Minor cleanups and rename to RenderManager

* It is no longer just a renderpass cache
* Revert variable name change from render_manager back to renderpass_cache

---------

+++ src/video_core/renderer_vulkan/vk_stream_buffer.h
@@ -35,10 +35,10 @@ public:
      * @param size Size to reserve.
      * @returns A pair of a raw memory pointer (with offset added), and the buffer offset
      */
-    std::tuple<u8*, u64, bool> Map(u64 size, u64 alignment);
+    std::tuple<u8*, u32, bool> Map(u32 size, u64 alignment);
 
     /// Ensures that "size" bytes of memory are available to the GPU, potentially recording a copy.
-    void Commit(u64 size);
+    void Commit(u32 size);
 
     vk::Buffer Handle() const noexcept {
         return buffer;
@@ -70,8 +70,8 @@ private:
     vk::BufferUsageFlags usage{};
     BufferType type;
 
-    u64 offset{};       ///< Buffer iterator.
-    u64 mapped_size{};  ///< Size reserved for the current copy.
+    u32 offset{};       ///< Buffer iterator.
+    u32 mapped_size{};  ///< Size reserved for the current copy.
     bool is_coherent{}; ///< True if the buffer is coherent
 
     std::vector<Watch> current_watches;           ///< Watches recorded in the current iteration.
